remove some async operations
