+++ src/hooks/useTrackingStorage/index.ts
@@ -1,9 +1,11 @@
 import { getTrackingStorageByProductId } from "./getTrackingStorageById";
 import { setTrackingStorage } from "./setTrackingStorage";
+import { syncTrackingStorage } from "./syncTrackingStorage";
 
 export function useTrackingStorage() {
   return {
     get: getTrackingStorageByProductId,
     set: setTrackingStorage,
+    sync: syncTrackingStorage,
   };
 }
